Start using Tracker.Search.Query.
This method has a horrible api and moreover vala dbus api is currently
not really capable of handling the return value of this method.
Nevertheless, it does improve performance a lot since now we get all the
metadata for all items at once rather than making a dbus call for each
item separately.

svn path=/trunk/; revision=576