core: Load config before setting log handler
Setting log handler before loading the config was suppressing all
debug messages. However now we have the problem that ALL messages
are printed even if they are of the level that is not allowed by the
config. Hopefully this problem is much more tolerable than the problem
this commit solves since there are not many messages printed before
config is loaded.