core: Limit the search results
This should improve performance of search in general a lot but it
depends on what slice in the search results the client is interested in.
If it asks for the last slice, there will be no performance improvements
to that search request because of this change. During an incremental
search (which all clients should be doing), the first few search results
will be extremely fast and we all know that this *will* effectively fool
the end-user. :)