core,ui,plugins: Work-around for automake bug
According to Nathan: It scrapes VALAFLAGS looking for options which
specify files, like -H or, significantly, --gir. Unfortunately, rather
than actually checking correctly, it does it in a weird "why did anybody
ever think that was the right thing to do?" kind of way. It has a list of
flags which specify files, and then if a flag is a substring of any of
those flags, it is considered to actually be one. So it mistakes "-g" for
"--gir" and assumes the following argument should be included when you
make dist. Until automake is fixed, you can fool it by putting -g at the
end of VALAFLAGS.