core: Pause pipeline when buffer grows too much
There is no way of knowing that an HTTP client has paused the playback but
we do know it when its not reading from socket anymore. So now we keep
track of number of buffers queued to be sent to the client and if this
number goes beyond a fixed amount, we pause the gstreamer pipeline until
client catches up and there is only a few number of queued buffers left.