core: Re-mux only if transcoding is not needed
MPEG2TS and WMV transcoders now bypass encoders if the stream is already
encoded in the needed codec. However, this also implies skipping the raw
audio and video tansformations and that in turn implies we are not exactly
being honest about the advertized DLNA profile name of the remuxed stream.
In most cases that shouldn't be a problem and this change should only make
users happy (hopefully).