core: SimpleContainer implements find_object()
The default find_object() implementation in base MediaContainer translates
find_object() to a search() query and since search() is recursive, the
whole tree under the container then gets a search() called on them so if
any of them has implemented find_object(), that implementation will never
be actually used. This is exactly what was happening in case of tracker
plugin and that was the main reason for it being so extremely slow in the
past few months.

Since most containers are SimpleContainer, this solves the problem for most
(if not all) of the plugins.