tracker: Join item ID components by ',' rather than ':'
Use ',' intead of ':' to join SearchContainer's ID and item's URN in the
child item's ID. This fixes item queries for Tracker backend. This issue
has been there since commit 51e4fdb (Feb 19 2010) but it didn't realize
so far because the default slow search used to kick-in and save the day.