tracker: Helper for generating search queries
Generating queries involves converting properties to key chains, adding
extra filter conditions required for a container and for mapping search
expressions. Once the results are retrieved, they need to be converted to
media objects. These tasks are performed by the new helper class.

It also provides a much more complete translation of Rygel's search
expression to SPARQL query. An intended side-affect of that is that the
searchable containers using this classes will not have to revert to the
default (potentially very very) slow search for most cases.