core,data: Correct meaning for 'upnp_enabled' config
'upnp_enabled' in the configuration now really means what it says:
Enable/Disable the UPnP devices. So its now set to 'true' by default and
if you set this to 'false', you can still stream media (currently only over
HTTP) from it. Of course, that means you need to know the URI of the item.