core,tracker: MetadataValues.is_child_id_unique -> SimpleContainer
Move is_child_id_unique() method of MetadataValues to parent
SimpleContainer and make it 'public' as its useful for all users.

This method was broken for (intially) empty containers after the
introduction of SimpleContainer.add_child_container. This patch also fixes
that issue by comparing the ID string in question with that of emtpy child
containers as well.