core: Don't ask Gst to handle HTTPByteSeek.stop
- Seems pipeline takes a very long time to playing state if we set the
  stop bit.
- HTTPGstSink doesn't let more bytes flow than requested anyways.

An issue in here is that pipeline will keep rolling now even after the
requested byte segment has been sent to the client. This issue however
should not realize since every decent client will abort the request
immediately after it has been satisfied and we stop/destroy the pipeline
when that happens.