core: Work-around standard issues for XBox
The MediaReceiverRegistrar uses an illegal character in the domain
name which makes UPnP tests fail. The XBox doesn't work without the
".", so we put the dash to make the UPnP test happy and replace it
with a dot when the XBox requests our service description.

Fixes (partially) UPnP CTT tests UDA-1.1.1, UDA-1.1.2, UDA-1.1.3
and UDA-1.1.4.