core: call remove_item for all MediaItems
Currently rygel-item-destroyer.remove_object explicitly removes
place_holder items from Tracker store. For actual media items that
have an associated file, it only removes the file and relies on
Tracker to handle the store update.

However, if the directory of that file is not monitored by Tracker,
the item never gets removed. This commit changes the behavior in
a way that Tracker store items are always explicitly removed, and
for non-placeholder items also the related file is removed. Also makes
Rygel satisfy DLNA guideline 7.3.139.1.