core: Make rygel single-instance.
The new instance steals the DBus name from the old instance which then
shuts itself down. This is a temporary solution to avoid a race
condition when multiple instances of Rygel being started by accident
are trying to read and write the SCDP files simultaneously, resulting
in empty SCDP files on-disk.