core: Restart find_object on container update
Before asynchronously descending in the hierarchy, we connect to the
updated signal of this container. If the container is updated, it might
be that this.children changed while we're iterating over it. If that
happens, we restart the foreach loop, but only 10 times. If we still
are interrupted, we give up.