media-export: Deactivate on slow Tracker start-up
If the startup of the Tracker plug-in is delayed (e.g. due to DBus
activation) the old method of disabling MediaExport did not work since
the Tracker plug-in was not active when MediaExport was loaded. The
plug-in now listens to notify::active of the Tracker plug-in and
deactivates itself once that changes to "true".

https://bugzilla.gnome.org/show_bug.cgi?id=662031