ui: Improve preferences dialog, bug 663631
Add tooltips to the widgets in the preferences dialog. Add +/-
ToolButtons for adding and removing shared directories. Update the
sensitivity of the remove toolbutton, based on whether a row in the URI
model is selected, or if the model is empty. Remove the GtkAlignments
and use padding on containers. Set the correct response ID on the close
button. Set the entry text column in the GtkBuilder file.