core: Fix simple search for offset > 0
On a search request where offset was > 0 and limit was 0 due to
limiting the result set to offset the result was always a set of one.

Also total_matches was set to 0 in this case although we know the
number of total matches since using a start offset doesn't limit the
results.

https://bugzilla.gnome.org/show_bug.cgi?id=653120