core: Do not bail out on missing thumbnail dir
This fixed the situation where you not yet have a ".thumbnails"
directory in your $HOME but have tumbler available. To do this, the code
now uses ListActivateableServices on D-Bus to check for the Thumbnailer1
interface and keeps on queuing incoming thumbnail requests until the
thumbnailer is either ready or not found. The queue will be passed on to
tumbler in the first case or flushed in the latter, marking the
thumbnailer not available then.

Unfortunately we cannot rely on Bus.get_proxy_sync to fail if it is not
available because it simply doesn't, looking at the gdbus code.

We stop handling the thumbnail directory ourselves and use GLib's
FileInfo for that