media-export: Ignore invalid colorDepth
media-export calls the GStreamer function gst_discoverer_video_info_get_depth
to determine the color depth of images and videos.  When this function fails
it returns 0.  The media-export plugin is not currently testing for this error
case and is treating 0 as a valid color depth.  Rygel in turn is reporting
a color depth of 0 for some files.  This seems to happen for all JPEGs on
my machine.  This patch fixes the problem by preventing Rygel from returning
a colorDepth resource for items whose color depth cannot be determined by
gst_discoverer_video_info_get_depth.

https://bugzilla.gnome.org/show_bug.cgi?id=684868