Fix build with non-leaky gupnp pkg-config files
The build system previously relied on the gupnp stack
leaking linker flags. Rygel should link to everything it
uses directly, not rely on anything else pulling in the
needed linker flags.

The gupnp stack is being fixed up to not leak linker flags
which will cause build failures if rygel is not fixed.

https://bugzilla.gnome.org/show_bug.cgi?id=685478