renderer: Add GetCurrentTransportActions method.
The GetCurrentTransportActions method and the CurrentTransportActions state
variable have been added to the AVTransport interface.  This method and its
related state variable need to be supported by renderers that support Seek
to be DLNA compliant.

Slightly modified to match current code. Note that this breaks the UPnP test
due to X_DLNA_SeekTime capability.

Fixes:
    https://bugzilla.gnome.org/show_bug.cgi?id=686638