build: core and server needs to link to gmodule and libxml.
Debian packaging tools warns about used symbols that are not available
in any of the linked libraries. These would probably all be pulled in
by users of librygel-core and librygel-server, but to be on the safe
side the binaries should ofcourse link to all libraries it uses
symbols from.

Sent by Andreas Henriksson.