server,media-export: Add and use signals emitted after tracking child changes.
child_added and child_removed are emitted when respectively
add_child_tracked and remove_child_tracked are finished. We use those
signals in media-export to store the container and, optionally, its
new child in the database. That simplifies a code a bit in
media-export and makes sure that update IDs are stored in database as
well.

https://bugzilla.gnome.org/show_bug.cgi?id=693577