server: Use custom equal function for updated_containers array.
Default equal function was simply comparing pointers, which is not
enough in this case. MediaExport plugin creates temporary
MediaContainer objects being virtual containers, so it is possible to
happen that we store such container in updated_containers array and
then later receive another one with the same id. Obviously new object
has different address than old one, so it was inserted into array as
well. It ended up with ContainerUpdateIDs having several entries with
the same id.

https://bugzilla.gnome.org/show_bug.cgi?id=694019