media-export: Prefer titles set by UPnP actions over discovered ones.
For this to work we have to mark an object created with CreateObject
call as guarded. It means that some of its properties (for now title
only) can't be overriden with a value from metadata extractor. The
only way to change a title of such object is by UpdateObject call.

All harvested objects that did not exist in database are not marked as
guarded. Doing UpdateObject on any object marks it as such though.

https://bugzilla.gnome.org/show_bug.cgi?id=683926