Only modify the description file from DescriptionFile
RootDeviceFactory used to modify the description xml directly. This is
not optimal as DescriptionFile now keeps the elements in the correct
order.

Move the icon element and service element creation code to
DescriptionFile, start using DescriptionFile public methods to modify
the description from RootDeviceFactory. Swap the order of ControlURL
and EventSubURL elements to match the specification.

https://bugzilla.gnome.org/show_bug.cgi?id=702451