server: Move knowledge of supported profiles
Query the supported profiles (all, upload) from the MediaServerPlugin,
provide a default implementation that does return everything from the
profiles supported by the MediaEngine.

Note: This is done via properties in this slightly complicated manner to not
break ABI this late in the cycle. Ideally it should be abstract/virtual
properties or abstract/virtual functions.