server: No OCM flags checking for CreateReference
There's no DLNA guidelines for reference creation, however if a container only
supports reference creation such as MediaExport's playlist containers, it
mustn't have the UPLOAD OCM flag set.

So we just check for writable container and let the creation fail later on in
the container if it's not allowed to have references there.

https://bugzilla.gnome.org/show_bug.cgi?id=707065