librygel-renderer: Allow underlying player to notify on volume changes
Currently there is no notification support on volume.
The last set value of the volume (via upnp) is cached in
rygel-rendering-control. But if the volume of the underlying renderer can
be changed in another way, for instance on a local UI. Such change is not
picked up by rygel-rendering-control and any control point would not be
aware of the volume change.

This patch adds notification support for volume.

Signed-off-by: Richard Röjfors <richard.rojfors@gmail.com>