Add support for pretty hostnames
Read and parse /etc/machine-info to obtain a pretty human readable
hostname (eg. "Bob's Laptop" instead of "bob-laptop"). If the
file is not found, or it can't be parsed, we fall back on the
normal hostname.

In GNOME, the pretty hostname can be configured in the control
center details and sharing panels. In other systems, it can be
changed using "hostnamectl --pretty".

https://bugzilla.gnome.org/show_bug.cgi?id=724138