renderer: Turn PlayerController into an interface
There is a need to be able to reimplement the rygel-player-controller
By changing the class to become an interface, other classes can
implement it.
The current implementation is moved to a default controller.
This controller is instantiated by get_controller of the plugin as
before, but now the get_controller function is defined virtual.
Making it possible for plugins to reimplement and return another
controller.

Signed-off-by: Richard Röjfors <richard.rojfors@gmail.com>

https://bugzilla.gnome.org/show_bug.cgi?id=732057