ruih: Make ServiceManager a singleton
Service was keeping information that belonged to ServiceManager and
basically constructing it in its static constructor to create a single
instance that was watching the file.

This commit moves all static members of Service to ServiceManager,
making it a full GObject and using the general singleton pattern we use
in Rygel.

This also fixes an potential issue in multi-homed hosts that the
static file-monitor would cancelled by Service's instance destructor if
one network interface went down.

Signed-off-by: Jens Georg <mail@jensge.org>