renderer Fix referencing of invalid GUPnP.Service
If AVTransport or RenderingControl is destroyed while a timeout is running,
ChangeLog is not destroyed with the Service and hence has a dangling service
pointer.

Rygel.ChangeLog was holding an unowned reference to GUPnP.Service. In a case
when the Service was already destroyed and ChangeLog's timeout method was
invoked, this resulted in accessing of already-destroyed instance. This patch
fixes it by using a WeakRef to the Service.

Signed-off-by: Milan Plzik <milan.plzik@streamunlimited.com>