Avoid always re-reading all mime-types on startup
This commits optimizes startup time, when scanning directories with
many files which are already cached and up-to-date. Before, we were
always opening and reading from them to determine their mime-type,
even though it was already in the local database. Now we re-use
this information, getting a speed up which can be even of
several minutes, especially on systems with rotational drives.

https://bugzilla.gnome.org/show_bug.cgi?id=778778