Allow UI build without gstreamer-video
If EXAMPLE_UI dependencies are not met, only display a warning instead
of ending configure with an error.
Indeed, currently, we can't build rygel UI without gstreamer-video which
is only a dependency of fullscreen-renderer application.

fullscreen-renderer won't be build without gstreamer as HAVE_GSTREAMER
has to be true to check HAVE_UI in examples/Makefile.am

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>